Conversion des date_xxx en guint (comme top) au lieu de time_t
git-svn-id: https://svn.abls-habitat.fr/repo/Watchdog@1553 9f22af27-4b39-4cfe-9f85-4bed9fef4693

+++ Watchdogd/Rs485/admin_rs485.c
@@ -60,9 +60,9 @@
                    module->rs485.e_min, module->rs485.e_max,
                    module->rs485.s_min, module->rs485.s_max, module->rs485.sa_min, module->rs485.sa_max,
                    (module->started      ? "TRUE " : "FALSE"),
-                   (Partage->top - (gint)module->date_requete)/10,
-                   (module->date_retente ? (Partage->top - (gint)module->date_retente)/10 : 0),
-                   (Partage->top - (gint)module->date_next_get_ana)/10
+                   (Partage->top - module->date_requete)/10,
+                   (module->date_retente ? (Partage->top - module->date_retente)/10 : 0),
+                   (Partage->top - module->date_next_get_ana)/10
                  );
        Write_admin ( client->connexion, chaine );
        liste_modules = liste_modules->next;
